Add user validation for habits

+++ src/db/habit-db.ts
@@ -1,20 +1,39 @@
 import prisma from '@/lib/db';
+import { getCurrentUser } from '@/utils/auth/current-user';
 
 import { Prisma } from '../../generated/prisma';
-export async function addHabit(data: Prisma.HabitCreateInput) {
+
+export async function addHabit(data: Prisma.HabitCreateArgs['data']) {
+  const user = await getCurrentUser();
+  if (!user) {
+    throw new Error('User not found');
+  }
+
   return prisma.habit.create({
     data,
   });
 }
 
-export async function editHabit(id: string, data: Prisma.HabitUpdateInput) {
+export async function editHabit(
+  id: string,
+  data: Prisma.HabitUpdateArgs['data']
+) {
+  const user = await getCurrentUser();
+  if (!user) {
+    throw new Error('User not found');
+  }
+
   return prisma.habit.update({
-    where: { id },
+    where: { id, userId: user.id },
     data,
   });
 }
 
 export async function getHabits() {
+  const user = await getCurrentUser();
+  if (!user) {
+    throw new Error('User not found');
+  }
   return prisma.habit.findMany({
     orderBy: {
       createdAt: 'desc',
@@ -22,16 +41,29 @@ export async function getHabits() {
     include: {
       completions: true,
     },
+    where: {
+      userId: user.id,
+    },
   });
 }
 
 export async function getHabitById(id: string) {
+  const user = await getCurrentUser();
+  if (!user) {
+    throw new Error('User not found');
+  }
+
   return prisma.habit.findUnique({
-    where: { id },
+    where: { id, userId: user.id },
   });
 }
 
 export async function removeHabit(id: string) {
+  const user = await getCurrentUser();
+  if (!user) {
+    throw new Error('User not found');
+  }
+
   return prisma.$transaction([
     prisma.habitCompletion.deleteMany({
       where: { habitId: id },
